requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
langchain==0.0.352
openai==1.3.5
pymysql==1.1.0
sqlalchemy==2.0.23
pydantic==2.5.2
bcrypt==4.0.1
pyjwt==2.8.0
pdfminer.six==20221105
faiss-cpu==1.7.4 